requirements.txt
Flask==3.1.2
Flask-Cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
SQLAlchemy==2.0.46
alembic==1.18.1
argon2-cffi==25.1.0
bleach==6.3.0
email-validator==2.3.0
gunicorn==24.1.1
mariadb==1.1.14
python-dotenv==1.2.1
redis==7.2.0
